+++ GameObjects/GameObjects/Influences/InfluenceKindPack/InfluenceKind3100.cs
@@ -6,7 +6,7 @@
 
     internal class InfluenceKind3100 : InfluenceKind
     {
-        private int multiple = 1;
+        private float multiple = 1;
 
         public override void ApplyInfluenceKind(Architecture architecture)
         {
@@ -17,7 +17,7 @@
         {
             try
             {
-                this.multiple = int.Parse(parameter);
+                this.multiple = float.Parse(parameter);
             }
             catch
             {
